# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/movie/findAll":{"get":{"operationId":"MovieController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Movie"],"security":[{"bearer":[]}]}},"/movie/FindOne{id}":{"get":{"operationId":"MovieController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Movie"],"security":[{"bearer":[]}]}},"/movie/Update{id}":{"patch":{"operationId":"MovieController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMovieDto"}}}},"responses":{"200":{"description":""}},"tags":["Movie"],"security":[{"bearer":[]}]}},"/movie/Remove{id}":{"delete":{"operationId":"MovieController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Movie"],"security":[{"bearer":[]}]}},"/movie/create":{"post":{"operationId":"MovieController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMovieDto"}}}},"responses":{"201":{"description":""}},"tags":["Movie"],"security":[{"bearer":[]}]}},"/user/findAll":{"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["User"]}},"/user/FindOne{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"],"security":[{"bearer":[]}]}},"/user/Update{id}":{"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["User"],"security":[{"bearer":[]}]}},"/user/Remove{id}":{"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"],"security":[{"bearer":[]}]}},"/user/register":{"post":{"operationId":"UserController_register","summary":"","description":"This endpoint is used for registering a new user.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["User"]}},"/genre/findAll":{"get":{"operationId":"GenreController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Genre"],"security":[{"bearer":[]}]}},"/genre/FindOne{id}":{"get":{"operationId":"GenreController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Genre"],"security":[{"bearer":[]}]}},"/genre/Update{id}":{"patch":{"operationId":"GenreController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateGenreDto"}}}},"responses":{"200":{"description":""}},"tags":["Genre"],"security":[{"bearer":[]}]}},"/genre/Remove{id}":{"delete":{"operationId":"GenreController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Genre"],"security":[{"bearer":[]}]}},"/genre/create":{"post":{"operationId":"GenreController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGenreDto"}}}},"responses":{"201":{"description":""}},"tags":["Genre"]}},"/director":{"post":{"operationId":"DirectorController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDirectorDto"}}}},"responses":{"201":{"description":""}},"tags":["Director"],"security":[{"bearer":[]}]},"get":{"operationId":"DirectorController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Director"],"security":[{"bearer":[]}]}},"/director/{id}":{"get":{"operationId":"DirectorController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Director"],"security":[{"bearer":[]}]},"patch":{"operationId":"DirectorController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDirectorDto"}}}},"responses":{"200":{"description":""}},"tags":["Director"],"security":[{"bearer":[]}]},"delete":{"operationId":"DirectorController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Director"],"security":[{"bearer":[]}]}},"/login":{"post":{"operationId":"LoginController_login","summary":"Log in a user.","parameters":[],"requestBody":{"required":true,"description":"The user's credentials.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginBodyDto"}}}},"responses":{"200":{"description":"The user has been successfully logged in."},"401":{"description":"Unauthorized."}},"tags":["Login"]}}},"info":{"title":"Lucas - Movie Catalog API Documentation","description":"\n    Este é um sistema de autenticação JWT com uma CRUD de um catálogo de filmes. Todos os endpoints desta CRUD só podem ser consumidos por um usuário autenticado.\n\n    Ferramentas utilizadas:\n    - TypeScript\n    - Nest.js\n    - TypeORM\n    - Swagger\n    - Docker\n    - Redis\n    - PostgreSQL\n\n    A arquitetura é composta por uma aplicação que fornece uma API RESTful em JSON, utilizando Redis como cache.\n\n    OBS: Todas as informações tratadas em cada endpoint são validadas.\n\n    Back-end:\n    A engenharia de qualidade é saber utilizar a ferramenta certa para o trabalho certo e seguir aprendendo constantemente sobre ela. Levando isso em consideração, a experiência com cada uma das ferramentas utilizadas é mencionada no README.\n\n    Deploy:\n    O deploy foi realizado no local mais confortável (exemplo: Amazon EC2, Heroku, Netlify, Google AppEngine, etc).\n\n    não consegui fazer tudo que eu queria, esse sistema foi feito em 24h\n  ","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"access-token":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"UpdateMovieDto":{"type":"object","properties":{}},"Director":{"type":"object","properties":{"id":{"type":"string","example":"6ef05f4e-49e1-4d14-8f06-682a9ba8adf3"},"createdAt":{"format":"date-time","type":"string","example":"2023/08/01 11:14"},"updatedAt":{"format":"date-time","type":"string","example":"2023/08/01 11:14"},"name":{"type":"string","example":"Christopher Nolan"},"movies":{"type":"array","items":{"$ref":"#/components/schemas/Movie"}}},"required":["id","name"]},"Movie":{"type":"object","properties":{"id":{"type":"string","example":"6ef05f4e-49e1-4d14-8f06-682a9ba8adf3"},"createdAt":{"format":"date-time","type":"string","example":"2023/08/01 11:14"},"updatedAt":{"format":"date-time","type":"string","example":"2023/08/01 11:14"},"title":{"type":"string","example":"The Shawshank Redemption"},"description":{"type":"string","example":"Two imprisoned men bond over a number of years..."},"releaseDate":{"format":"date-time","type":"string","example":"1994-09-22"},"duration":{"type":"number","example":142},"rating":{"type":"number","example":9.3},"coverImageUrl":{"type":"string"},"genres":{"type":"array","items":{"$ref":"#/components/schemas/Genre"}},"directors":{"type":"array","items":{"$ref":"#/components/schemas/Director"}}},"required":["id","title","description","releaseDate","duration","rating"]},"Genre":{"type":"object","properties":{"id":{"type":"string","example":"6ef05f4e-49e1-4d14-8f06-682a9ba8adf3"},"createdAt":{"format":"date-time","type":"string","example":"2023/08/01 11:14"},"updatedAt":{"format":"date-time","type":"string","example":"2023/08/01 11:14"},"name":{"type":"string","example":"Drama"},"movies":{"type":"array","items":{"$ref":"#/components/schemas/Movie"}}},"required":["id","name"]},"CreateMovieDto":{"type":"object","properties":{"title":{"type":"string","example":"The Shawshank Redemption","description":"Title of the movie"},"description":{"type":"string","example":"Two imprisoned men bond over a number of years...","description":"Description of the movie"},"releaseDate":{"format":"date-time","type":"string","example":"1994-09-22","description":"Release date of the movie"},"duration":{"type":"number","example":142,"description":"Duration of the movie"},"rating":{"type":"number","example":9.3,"description":"Rating of the movie"},"coverImageUrl":{"type":"string","description":"URL of the movie cover image"},"genres":{"type":"array","items":{"$ref":"#/components/schemas/Genre"}},"directors":{"type":"array","items":{"$ref":"#/components/schemas/Director"}},"genreIds":{"description":"IDs of genres associated with the movie","type":"array","items":{"type":"string"}},"directorIds":{"description":"IDs of directors associated with the movie","type":"array","items":{"type":"string"}}},"required":["title","description","releaseDate","duration","rating","coverImageUrl","genreIds","directorIds"]},"UpdateUserDto":{"type":"object","properties":{}},"CreateUserDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com","description":"Email of the user"},"password":{"type":"string","example":"password","description":"Password of the user"},"avatarUrl":{"type":"string","description":"URL of the user avatar"},"isValidUser":{"type":"boolean","description":"Is user valid"},"active":{"type":"boolean","example":true,"description":"Is user active"},"isAdmin":{"type":"boolean","example":false,"description":"Is user an admin"}},"required":["email","password"]},"UpdateGenreDto":{"type":"object","properties":{}},"CreateGenreDto":{"type":"object","properties":{"name":{"type":"string","example":"Drama"}}},"CreateDirectorDto":{"type":"object","properties":{"name":{"type":"string","example":"Christopher Nolan"}}},"UpdateDirectorDto":{"type":"object","properties":{}},"LoginBodyDto":{"type":"object","properties":{"email":{"type":"string","description":"The email of the user."},"password":{"type":"string","description":"The password of the user."}},"required":["email","password"]}}}}